Split settings and opts to separate mods
